cse-chatbot/requirements.txt
Flask
python-dotenv
openai
google-generativeai
